php pretty much done with...
the only thing not implemented is the logging of educational information
readings performed by the user. All worked on the server but should
undergo some kind of testing. A few tweaks will be needed regarding how
we control the flow of the game i.e. redirects between the various
pages. Restructuring of these will affect the php control flow also.